Fixed new blog entry and file name fixes
There are three primary changes in this commit.  The first is to correct
the name of the GitHub pages blog.  The second is the addition of the
configuration for MDWiki (including Caddy).  The third change is a
modification to the posts template. The other change is an update to the
Debian Dotfiles configuration entry.

The Debian Dotfiles configuration needs to be updated to include the
configuration of VIm and i3w.